// pdfix-sdk-example-angular/src/app/interfaces/pv-rendered-page-div-props.ts
/**
 * Interface for sample rendered page div
 * @interface PvRenderedPageDivProps
 */
export interface PvRenderedPageDivProps {
    /** Width of the segment to be rendered. */
    width: number;
    /** Height of the segment to be rendered. */
    height: number;
    /** Base64 encoded image data of the segment. */
    base64: string;
}